+++ scripts/lib/color-scheme/index.ts
@@ -1,0 +1,1 @@
+export * from './storage.ts';
