+++ scripts/lib/color-scheme/models.ts
@@ -1,0 +1,6 @@
+export const SchemeStatus = {
+  System: 'system',
+  Light: 'light',
+  Dark: 'dark',
+} as const;
+export type SchemeStatus = (typeof SchemeStatus)[keyof typeof SchemeStatus];
