+++ scripts/lib/env.d.ts
