+++ scripts/lib/meta/data-models.ts
@@ -1,0 +1,4 @@
+export interface PageImage {
+  src: string;
+  alt: string;
+}
