+++ scripts/lib/meta/index.ts
@@ -1,2 +1,6 @@
-export * from './zod-models.ts';
+export * from './frontmatter-models.ts';
+export * from './frontmatter.ts';
 export { getLangaugeLabel } from './lang.ts';
+export * from './page.ts';
+export * from './resource.ts';
+export * from './data-models.ts';
