+++ scripts/lib/schema/index.ts
@@ -1,0 +1,2 @@
+export * from './schema.ts';
+export * from './schema-models.ts';
