+++ src/components/Datetime.astro
@@ -1,22 +1,12 @@
 ---
+import { formatDate } from "@lib/meta";
+
 export interface Props {
   datetime: Date | undefined;
   title?: string;
 }
 
 const { datetime, title = "Datetime", ...props } = Astro.props;
-
-function formatDate(dateObject: Date) {
-  if (typeof dateObject === "string") {
-    dateObject = new Date(dateObject)
-  }
-
-  const year = String(dateObject.getFullYear()).padStart(4, "0");
-  const month = String(dateObject.getMonth() + 1).padStart(2, "0");
-  const date = String(dateObject.getDate()).padStart(2, "0");
-
-  return `${year}-${month}-${date}`;
-}
 ---
 
 {
