+++ src/components/meta/Meta.astro
@@ -1,6 +1,5 @@
 ---
-import { Website, Page } from "@lib/meta";
-import { ThemeColor } from "@lib/metagg";
+import { Website, Page, ThemeColor } from "@lib/meta";
 import { SchemaProps } from "@lib/schema";
 
 import Settings from "@components/meta/Settings.astro";
