+++ src/components/meta/OpenGraph.astro
@@ -3,8 +3,7 @@
   https://ogp.me/
   https://developers.google.com/search/docs/appearance/structured-data/intro-structured-data
 */
-
-import { PageImage, getFormat } from "@lib/metagg";
+import { getFormat, PageImage } from "@lib/meta";
 
 export interface Props {
   title: string;
