+++ src/components/meta/PageData.astro
@@ -1,5 +1,5 @@
 ---
-import { PageImage } from "@lib/metagg";
+import { PageImage } from "@lib/meta";
 
 export interface Props {
   title: string;
