+++ src/components/meta/PreAPI.astro
@@ -1,6 +1,6 @@
 ---
 import { createGoogleFontURL } from "@lib/font";
-import { Resources } from "@lib/metagg";
+import { Resources } from "@lib/meta";
 
 import PreConnect from "@components/meta/PreConnect.astro";
