+++ src/components/meta/Settings.astro
@@ -4,7 +4,7 @@
  * This is also the case with dark themes only.
  */
 
-import { ThemeColor } from "@lib/metagg";
+import { ThemeColor } from "@lib/meta";
 
 export interface Props {
   noIndex: boolean;
