+++ src/components/meta/TwitterCards.astro
@@ -4,7 +4,7 @@
   https://developer.twitter.com/en/docs/twitter-for-websites/cards/overview/summary-card-with-large-image
 */
 
-import { PageImage } from "@lib/metagg";
+import { PageImage } from "@lib/meta";
 
 export interface Props {
   title: string;
