+++ src/components/sidebar/HeadingList.svelte
@@ -2,7 +2,7 @@
   import { onMount } from 'svelte';
   import { writable } from 'svelte/store';
 
-  import type { Headings } from '@lib/metagg';
+  import type { Headings } from '@lib/meta';
 
   export let isDev = false;
   /** Value automatically generated by astro. */
