+++ src/layouts/BaseLayout.astro
@@ -1,6 +1,5 @@
 ---
-import { Website, Page, Frontmatter } from "@lib/meta";
-import { toPage, getPersons, ThemeColor, Resources } from "@lib/metagg";
+import { Website, Page, Frontmatter, toPage, getPersons, ThemeColor, Resources } from "@lib/meta";
 import { articleSchema, Schemas, SchemaProps, webSiteSchema } from "@lib/schema";
 
 import PreAPI from "@components/meta/PreAPI.astro";
