+++ src/layouts/BlogLayout.astro
@@ -1,6 +1,5 @@
 ---
-import { Frontmatter } from "@lib/meta";
-import {  Headings } from "@lib/metagg";
+import { Frontmatter, Headings } from "@lib/meta";
 import { getFont } from "@lib/font";
 
 import BaseLayout from "@layouts/BaseLayout.astro";
