+++ src/layouts/MainLayout.astro
@@ -3,7 +3,7 @@ import { GetImageResult } from "astro";
 import { getImage } from "astro:assets";
 
 import { getFont } from "@lib/font";
-import { Backgrounds } from "@lib/metagg";
+import { Backgrounds } from "@lib/meta";
 
 import BaseLayout from "@layouts/BaseLayout.astro";
 import Header from "@layouts/Header.astro";
