+++ src/layouts/Sidebar.astro
@@ -1,6 +1,5 @@
 ---
-import { Frontmatter } from "@lib/meta";
-import { Headings } from "@lib/metagg";
+import { Frontmatter, Headings } from "@lib/meta";
 
 import HeadingList from "@components/sidebar/HeadingList.svelte";
 import ShareButtons from "@components/sidebar/ShareButtons.astro";
